Minor CLI code improvements
